+++ CHANGELOG.md
@@ -2,4 +2,4 @@
 * Fixed conditional items not showing in creative tabs
 * Fixed Squeeze Trees not generating
 * Added Scorched retrogen
-* Added a config screen (incomplete)
+* Added a config screen